添加固定signal向任务signal转化的questnode
添加必须由parm指定Quest的Incident

修复BulletBase地图Map被提前销毁的bug

添加可主动进入地图的MapParent

优化Quest和WorldObject的关联

整理文件夹结构

+++ WorldObject/SaleRequestComp.cs
@@ -28,7 +28,7 @@ public class SaleRequestComp : WorldObjectComp
     {
         if (ActiveRequest)
         {
-            return "CaravanRequestInfo".Translate(TradeRequestUtility.RequestedThingLabel(saleRQ_ThingDef, saleRQ_Count).CapitalizeFirst(), (saleRQ_Expiration - Find.TickManager.TicksGame).ToStringTicksToDays(), (saleRQ_ThingDef.GetStatValueAbstract(StatDefOf.MarketValue) * (float)saleRQ_Count).ToStringMoney());
+            return "CaravanRequestInfo".Translate(TradeRequestUtility.RequestedThingLabel(saleRQ_ThingDef, saleRQ_Count).CapitalizeFirst(), (saleRQ_Expiration - Find.TickManager.TicksGame).ToStringTicksToDays(), (saleRQ_ThingDef.GetStatValueAbstract(StatDefOf.MarketValue) * saleRQ_Count).ToStringMoney());
         }
         return null;
     }
